add Feature_PlayerCollision
this uses a capsule for allow for detection of player collision. In default settings
it works automatically with the Locomotion_Stick featurea and also with the
Feature_Falling in automatic settings. But specific behaviour will need (as always)
changeing the code and adapting it to the needs of the project.