PC Oculus should bet set to 90Hz
Oculus on PC will respect target_fps as a maximum, but it will automatically turn down the fps to what ever the HMD requires.
Thus setting this to the maximum of 90Hz will allow those HMDs to run at their proper rate, and Oculus will handle reducing it for other HMDs that need a lower rate.